feat(kernel): GC overhaul phases 1-4 — binary search, pool resize, autorelease roots, free lists
Phase 1: Replace O(n²) linear scan in gc_visit_id_ptr() GC_UPDATE
phase with binary search on the sorted relocs[] array. Reduces
update cost from O(n*m) to O(m*log n) where n=reloc_count and
m=root pointer count.

Phase 2: Double OBJC_INTERP_STRING_POOL_SIZE from 64KB to 128KB.
The 64KB limit was set when the kernel was smaller; 128KB is still
tiny by WASM standards and eliminates the most common OOM case.

Phase 3: Integrate autorelease pools with GC. Change
AutoreleasePool.object_markers from unsigned int[] to id[] so
gc_visit_roots() can visit them as roots. Fix [obj autorelease]
handler to store id directly instead of casting to unsigned int.
When a pool drains (pool_depth decrements), objects in that pool
are no longer roots and become eligible for collection.

Phase 4: Add free list infrastructure for blocks, invocations,
and associations (matching the existing coll_free_list pattern).
Add block_init(), invocation_init(), association_init() functions
and call them from both reset paths. Actual slot reuse will be
wired in during the handle-based object table migration.

+++ objc-jupyter-wasm/kernel/objc_interp_messages.c
@@ -743,9 +743,7 @@ Value parse_message_send(Parser *p) {
             if (g_ctx.pool_depth > 0) {
                 AutoreleasePool *pool = &g_ctx.pools[g_ctx.pool_depth - 1];
                 if (pool->count < MAX_AUTORELEASE_OBJECTS) {
-                    /* Store the object marker as a uintptr_t cast to unsigned int.
-                     * String pool pointers are already within 32-bit range. */
-                    pool->object_markers[pool->count++] = (unsigned int)(uintptr_t)receiver;
+                    pool->object_markers[pool->count++] = receiver;
                 }
             }
             return value_from_id(receiver);
